HIVE-24599: Refactor trim vector expressions
