+++ ql/src/java/org/apache/hadoop/hive/ql/exec/vector/expressions/StringTrimColCol.java
@@ -19,14 +19,11 @@
 package org.apache.hadoop.hive.ql.exec.vector.expressions;
 
 import org.apache.hadoop.hive.ql.exec.vector.BytesColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.VectorExpressionDescriptor;
-import org.apache.hadoop.hive.ql.exec.vector.VectorizedRowBatch;
-import org.apache.hadoop.hive.ql.metadata.HiveException;
 
 /**
  * Vectorized TRIM with a string column and a trim-characters column.
  */
-public class StringTrimColCol extends VectorExpression {
+public class StringTrimColCol extends StringTrimColColBase {
   private static final long serialVersionUID = 1L;
 
   public StringTrimColCol(int strCol, int trimCharsCol, int outputColumnNum) {
@@ -38,71 +35,10 @@ public class StringTrimColCol extends VectorExpression {
   }
 
   @Override
-  public void evaluate(VectorizedRowBatch batch) throws HiveException {
-    if (childExpressions != null) {
-      super.evaluateChildren(batch);
-    }
-
-    BytesColumnVector strV = (BytesColumnVector) batch.cols[inputColumnNum[0]];
-    BytesColumnVector trimV = (BytesColumnVector) batch.cols[inputColumnNum[1]];
-    BytesColumnVector outV = (BytesColumnVector) batch.cols[outputColumnNum];
-    int[] sel = batch.selected;
-    int n = batch.size;
-    if (n == 0) {
-      return;
-    }
-
-    outV.initBuffer();
-    NullUtil.propagateNullsColCol(strV, trimV, outV, sel, n, batch.selectedInUse);
-
-    if (strV.isRepeating && trimV.isRepeating) {
-      if (!outV.isNull[0]) {
-        applyTrim(outV, strV, 0, trimV, 0, 0);
-      }
-      outV.isRepeating = true;
-      return;
-    }
-
-    if (batch.selectedInUse) {
-      for (int j = 0; j != n; j++) {
-        int i = sel[j];
-        if (!outV.isNull[i]) {
-          applyTrim(outV, strV, i, trimV, trimV.isRepeating ? 0 : i, i);
-        }
-      }
-    } else {
-      for (int i = 0; i != n; i++) {
-        if (!outV.isNull[i]) {
-          applyTrim(outV, strV, i, trimV, trimV.isRepeating ? 0 : i, i);
-        }
-      }
-    }
-  }
-
-  private void applyTrim(BytesColumnVector outV, BytesColumnVector strV, int strIndex,
+  protected void func(BytesColumnVector outV, BytesColumnVector strV, int strIndex,
       BytesColumnVector trimV, int trimIndex, int outIndex) {
     StringTrimColScalarBase.trimBoth(outV, strV.vector[strIndex], strV.start[strIndex],
         strV.length[strIndex], trimV.vector[trimIndex], trimV.start[trimIndex],
         trimV.length[trimIndex], outIndex);
-  }
-
-  @Override
-  public String vectorExpressionParameters() {
-    return getColumnParamString(0, inputColumnNum[0]) + ", "
-        + getColumnParamString(1, inputColumnNum[1]);
-  }
-
-  @Override
-  public VectorExpressionDescriptor.Descriptor getDescriptor() {
-    return (new VectorExpressionDescriptor.Builder())
-        .setMode(VectorExpressionDescriptor.Mode.PROJECTION)
-        .setNumArguments(2)
-        .setArgumentTypes(
-            VectorExpressionDescriptor.ArgumentType.STRING_FAMILY,
-            VectorExpressionDescriptor.ArgumentType.STRING_FAMILY)
-        .setInputExpressionTypes(
-            VectorExpressionDescriptor.InputExpressionType.COLUMN,
-            VectorExpressionDescriptor.InputExpressionType.COLUMN)
-        .build();
   }
 }
